build: generate openssl config for BSD-x86
Add BSD-x86 to the list of arches and regenerate everything.

Everything in deps/openssl/config/archs/BSD-x86 is auto-generated,
everything else is manual edits to include the right files in the
right places.

I cheated a little: I didn't check in changes to files for other arches
because they contained mostly churn (updated buildstamps and such.)

Fixes: https://github.com/nodejs/node/issues/20326

PR-URL: https://github.com/nodejs/node/pull/28806
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>